requirements.txt
boto3 
python-dotenv
pandas
pylint
pytest
psycopg2
datetime
argparse
confluent_kafka
